requirements.txt
# Echo - Voice Controlled Command Runner
# Dependencies

# Audio processing
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.1

# Speech recognition
openai-whisper==20231117

# Text-to-speech
pyttsx3==2.90

# HTTP requests
requests==2.31.0

# Async support
asyncio-mqtt==0.13.0

# Local LLM support
ollama==0.1.3